Change param type to any
The WFE types the template data as `FormValues`, which is defined by
the third-party dep Formik. While we do expect an object at all times,
this might be typed differently in different applications and then
requires type casting. The validaiton functionality does not break
when passing in other types, they will simply return as not valid.